Moved scroll top button to ui folder
